Separate test that runs ert to own file
