+++ tests/test_functions.py
@@ -1,10 +1,8 @@
 """Test utility ecl2csv"""
 
 import os
-from pathlib import Path
 from numpy.ma import allclose, allequal
 from shutil import copytree
-from subprocess import PIPE, Popen
 from time import sleep
 from io import BytesIO
 import pandas as pd
@@ -70,42 +68,6 @@ def check_sumo(case_uuid, tag_prefix, correct, class_type, sumo):
     sleep(SLEEP_TIME)
 
     sumo.delete(path, query)
-
-
-def write_ert_config_and_run(runpath):
-    ert_config_path = "sim2sumo.ert"
-    encoding = "utf-8"
-    ert_full_config_path = runpath / ert_config_path
-    print(f"Running with path {ert_full_config_path}")
-    with open(ert_full_config_path, "w", encoding=encoding) as stream:
-
-        stream.write(
-            f"DEFINE <SUMO_ENV> dev\nNUM_REALIZATIONS 1\nMAX_SUBMIT 1\nRUNPATH {runpath}\nFORWARD_MODEL SIM2SUMO"
-        )
-    with Popen(
-        ["ert", "test_run", str(ert_full_config_path)],
-        stdout=PIPE,
-        stderr=PIPE,
-    ) as process:
-        stdout, stderr = process.communicate()
-
-    print(
-        f"After ert run all these files where found at runpath {[item.name for item in list(Path(runpath).glob('*'))]}"
-    )
-    if stdout:
-        print("stdout:", stdout.decode(encoding), sep="\n")
-    if stderr:
-        print("stderr:", stderr.decode(encoding), sep="\n")
-    try:
-        error_content = Path(runpath / "ERROR").read_text(encoding=encoding)
-    except FileNotFoundError:
-        error_content = ""
-    assert (
-        not error_content
-    ), f"ERROR file found with content:\n{error_content}"
-    assert Path(
-        runpath / "OK"
-    ).is_file(), f"running {ert_full_config_path}, No OK file"
 
 
 def check_expected_exports(expected_exports, shared_grid, prefix):
@@ -396,41 +358,6 @@ def test_convert_to_arrow():
     dframe["DATE"] = dframe["DATE"].astype("datetime64[ms]")
     table = convert_to_arrow(dframe)
     assert isinstance(table, pa.Table), "Did not convert to table"
-
-
-def test_sim2sumo_with_ert(scratch_files, case_uuid, sumo, monkeypatch):
-    monkeypatch.chdir(scratch_files[0])
-    real0 = scratch_files[0]
-    write_ert_config_and_run(real0)
-    expected_exports = 88
-    path = f"/objects('{case_uuid}')/search"
-    results = sumo.post(
-        path,
-        json={
-            "query": {
-                "bool": {
-                    "must_not": [
-                        {
-                            "terms": {
-                                "class.keyword": [
-                                    "case",
-                                    "iteration",
-                                    "realization",
-                                ]
-                            }
-                        }
-                    ],
-                }
-            },
-            "size": 0,
-            "track_total_hits": True,
-        },
-    ).json()
-
-    returned = results["hits"]["total"]["value"]
-    assert (
-        returned == expected_exports
-    ), f"Supposed to upload {expected_exports}, but actual were {returned}"
 
 
 @pytest.mark.parametrize("real,nrdfiles", [(REEK_REAL0, 2), (REEK_REAL1, 5)])
